requirements.txt
Flask
PyMuPDF
pdf2docx
PyPDF2
python-pptx
Pillow
